// src/@types/form.d.ts
export interface ContactFormData {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  company?: string;
  howDidYouHearAboutUs?: string;
  website?: boolean;
  mobileApp?: boolean;
  marketingAutomation?: boolean;
  augmentedReality?: boolean;
  emailMarketing?: boolean;
  branding?: boolean;
  socialMedia?: boolean;
  maketingOther?: boolean;
  wantHelp?: boolean;
  details?: string;
  newsletter?: boolean;
  "g-recaptcha-response"?: string;
}